Log In/Sign Up/ Interceptors
We created routing for the application, basic layouts like nav and footer, we also created reactives forms for log in and sign up functionality. We created an Interceptor so we can manipulate every response to the API and get a Custom error in a modal component. After that we set another interceptor that will put for every request the corresponding jwt token from local storage.