AuthStore and TvShowReminderDialog
On the auth store service we are decoding the token to get the user id and save it as a entry on local Storage. We can also get this user id from local storage. We did this, because we need the id of the logged user when we save a reminder. Also i created the html and funcionality of the form in the TvShowReminderComponent. Next step is to creat the service and save the tv show reminder information