Refactored Dialogs, Added funcionality for user tv show reminders, creating interfaces
First, we refactored the details dialog and the user tv show dialog to pass an object with relevant information so we can use it to create its reminders. Then we refactored the tv show reminders dialog component, when the user creates its own tv show, it will send it to the reminder dialog, save this tv show and after that it will create the reminder for the saved tv show. We also created the service so we can save the tv show