Refactored the way we save and update reminders. Also i add delete funcionality
First , we are watching from the tv show reminder component differents observables, one when the user save a reminder, when the user update a reminder and finally when he deletes a reminder. For that we refactored the communication service. We also refactored the way we add, update and delete the data source subject so we always have the latest information in the table.