Refactored communication betweem the  child dialogs and the table component
Instead of subscribing to different observables to know what operation i did with a reminder, I created a new object to know the operation (save-update-delete). So now I only subscribe to one observable in the reminders component, and after that i can check the operation and know what i have to do. (Push the reminder to the data source, update the data source or delete the data source in the data source service).