delete reminder funcionality
We refactored the way we call the service to delete a reminder. Now we are passing the index and size of the page, so we can delete a reminder and get the information of the page without this reminder. We then emmited this page, so the reminder component can get this information. After that, we can update the datasource with this page we retrieve.