Add and delete reminder to data source
I refactored the way the paginator works. Before i only had the load reminder funcionality for each pagination event. Now I am considering a change of pages when the user add a reminder or delete a reminder. We refactored the data source reminder service, to emit a object with the information we need so we can go to a specific page. Now every time the user add a reminder we will redirect to the last page so he can view the reminder that he added. The same when he deletes all the reminders of a page, we are going to load the previous page.