Make table rows clickable to display modals
Now when the user click a row from the reminders table, it will open the TvShowDetails dialog to view the details of the tv show. Or it will open the UserTvShow Dialog if the reminder is from a tv show created by the user. If thats the case, we are going to allow the user to update its tv show.