Modified Profile and Footer component. Also now we can  update a user tv show
Now the user can click the row of the table and open the user tv show dialog, so it can use it to update the user tv show. We update it in the API, and then we push this changes through the communication service, to have the state of the reminder with this new update. Also we added information to the profile component and footer component